+++ prisma/seeds/create-professional-user.ts
@@ -6,7 +6,7 @@ import {
   ProfessionalDocumentType,
   TypeOfQuery,
 } from '../.././src/infra/database/generated/client';
-import { hashPassword } from '../../src/core/shared/utils';
+import { hashPassword } from '../../src/utils';
 
 export const CreateProfessionalUser = async (prisma: PrismaClient) => {
   const specialties = await prisma.specialty.findMany();
